+++ RELEASE-NOTES.md
@@ -19,6 +19,7 @@
 ## Deprecations
 
 ## New additions
+* Added `--private-link` flag to `snow spcs image-registry url` command for retrieving private link URLs.
 
 ## Fixes and improvements